加了点数据为""校验和身份证校验

+++ src/main/java/com/healthcode/service/impl/StudentServiceImpl.java
@@ -1,5 +1,7 @@
 package com.healthcode.service.impl;
 
+import cn.hutool.Hutool;
+import cn.hutool.core.util.IdcardUtil;
 import com.healthcode.common.HealthCodeException;
 import com.healthcode.dao.ClazzDao;
 import com.healthcode.dao.StudentDailyCardDao;
@@ -10,13 +12,16 @@ import com.healthcode.domain.Student;
 import com.healthcode.domain.StudentDailyCard;
 import com.healthcode.dto.CurrentDailyCard;
 import com.healthcode.service.IStudentService;
+import com.healthcode.utils.CheckValueUtil;
 import com.healthcode.utils.ExcelUtil;
 import com.healthcode.utils.JudgeHealthCodeTypeUtil;
 import com.healthcode.vo.StudentDailyCardStatistic;
 import com.healthcode.vo.StudentDailyCardVo;
+import javafx.beans.binding.ObjectExpression;
 import org.jetbrains.annotations.Nullable;
 
 import javax.servlet.http.Part;
+import javax.swing.*;
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.Date;
@@ -35,6 +40,10 @@ public class StudentServiceImpl implements IStudentService {
 
     @Override
     public Student login(String username, String password) {
+        //校验数据
+        if (CheckValueUtil.checkStringHelper(username, password)){
+            throw new HealthCodeException("信息不可为空");
+        }
         //学生登录
         Student student = studentDao.getByUsername(username);
         if (Objects.isNull(student) || !student.getPassword().equals(password)) {
@@ -45,6 +54,10 @@ public class StudentServiceImpl implements IStudentService {
 
     @Override
     public Boolean checkStudentDailyCardToday(Student student) {
+        //校验数据
+        if (Objects.isNull(student)){
+            throw new HealthCodeException("当前学生不存在");
+        }
         //true:已打卡 false:未打卡
         return studentDailyCardDao.getToDayCardByStudentID(student.getId()) != null;
     }
@@ -56,7 +69,7 @@ public class StudentServiceImpl implements IStudentService {
             @Nullable Integer collegeId) {
         StudentDailyCardStatistic statistic = new StudentDailyCardStatistic();
         List<Student> students;
-        System.out.println("class:" + clazzId + " major:" + majorId + " college:" + collegeId);
+
         if (!Objects.isNull(clazzId)) {
             students = studentDao.listAllByClazzId(clazzId);
         } else if (!Objects.isNull(majorId)) {
@@ -115,6 +128,16 @@ public class StudentServiceImpl implements IStudentService {
 
     @Override
     public void insertStudent(String id, String name, Integer classId, String idCard) {
+        //校验数据
+        if (CheckValueUtil.checkStringHelper(id, name, idCard)){
+            throw new HealthCodeException("信息不可为空");
+        }
+        if (Objects.isNull(clazzDao.getById(classId))){
+            throw new HealthCodeException("该班级号不存在");
+        }
+        if (!IdcardUtil.isValidCard(idCard)){
+            throw new HealthCodeException("身份证无效");
+        }
         //添加学生
         //密码默认为身份证后六位
         studentDao.insert(id, name, idCard.substring(idCard.length() - 6), classId, idCard);
@@ -122,6 +145,10 @@ public class StudentServiceImpl implements IStudentService {
 
     @Override
     public byte[] showHealthCode(Student student) {
+        //校验数据
+        if (Objects.isNull(student)){
+            throw new HealthCodeException("信息不可为空");
+        }
         //返回学生二维码
         List<HealthCodeType> healthCodeTypeList = studentDailyCardDao.judgeHealthCodeTypeByPast(student.getId());
         HealthCodeType healthCodeType = JudgeHealthCodeTypeUtil.judgeHealthTypeHelper(healthCodeTypeList);
@@ -136,6 +163,10 @@ public class StudentServiceImpl implements IStudentService {
 
     @Override
     public void submitDailyCard(Student student, CurrentDailyCard card) {
+        //校验数据
+        if (Objects.isNull(student) || Objects.isNull(card)){
+            throw new HealthCodeException("信息不可为空");
+        }
         //判断当前健康状态
         HealthCodeType healthCodeType = qrCodeService.judgeQRCodeType(card);
         //提交至数据库
@@ -145,6 +176,10 @@ public class StudentServiceImpl implements IStudentService {
 
     @Override
     public void addStudentFromExcel(Part filePart) {
+        //校验数据
+        if (Objects.isNull(filePart)){
+            throw new HealthCodeException("导入的文件为空");
+        }
         //将Excel文件中的学生数据导入数据库
         try {
             List<Object> studentData = ExcelUtil.readLessThan1000RowBySheet(filePart.getInputStream(), null);
@@ -175,16 +210,41 @@ public class StudentServiceImpl implements IStudentService {
 
     @Override
     public void deleteById(String id) {
+        //校验数据
+        if (CheckValueUtil.checkStringHelper(id)){
+            throw new HealthCodeException("信息不可为空");
+        }
         studentDao.delete(id);
     }
 
     @Override
     public Student getById(String id) {
+        //校验数据
+        if (CheckValueUtil.checkStringHelper(id)){
+            throw new HealthCodeException("信息不可为空");
+        }
         return studentDao.getByUsername(id);
     }
 
     @Override
-    public void updateStudent(String id, Integer classId, String name, String idCard) {
-        //TODO
+    public void updateStudent(String id, String password, Integer classId, String name, String idCard) {
+        //校验数据
+        if (CheckValueUtil.checkStringHelper(id)){
+            throw new HealthCodeException("id不可为空");
+        }
+        //获取当前数据
+        Student student = studentDao.getByUsername(id);
+        if(!Objects.isNull(student)){
+            name = "".equals(name) ? student.getName() : name;
+            password = "".equals(password) ? student.getPassword() : password;
+            classId = Objects.isNull(classId) ? student.getClazz().getId() : classId;
+            if(!"".equals(idCard) && !IdcardUtil.isValidCard(idCard)){
+                throw new HealthCodeException("身份证无效");
+            }
+            idCard = "".equals(idCard) ? student.getIdCard() : idCard;
+            studentDao.alter(id, name, password, classId, idCard);
+        }else {
+            throw new HealthCodeException("当前学生不存在");
+        }
     }
 }
